requirements.txt
httpx
numpy
onnxruntime
openai
opuslib
pyaudio
pydub
PyYAML
scipy
silero-vad
